requirements.txt
# 游戏装备图像识别系统依赖包
# 核心图像处理库
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0

# 可选依赖（用于性能优化和扩展功能）
# opencv-contrib-python>=4.8.0  # 包含额外的OpenCV模块
# scikit-image>=0.19.0         # 高级图像处理功能
# matplotlib>=3.5.0            # 图像可视化和分析
# tqdm>=4.64.0                 # 进度条显示

# 开发和测试依赖
# pytest>=7.0.0               # 单元测试框架
# pytest-cov>=4.0.0           # 测试覆盖率
# black>=22.0.0                # 代码格式化
# flake8>=5.0.0                # 代码检查